backend/requirements.txt
Flask==3.0.3
stripe==10.8.0
pyjwt==2.9.0
flask-cors==4.0.1
python-dotenv==1.0.1
pymongo[srv]==4.5.0
pylint==3.2.6
bandit==1.7.9
google-cloud-storage==2.18.2
openai==1.42.0
Flask-Limiter==3.8.0
Authlib==1.2.1
gunicorn==23.0.0
gevent==24.2.1
Flask-Caching==2.3.0